Handle nil test.name (#69)
I assume the treesitter API changed (again) and caused this to start
breaking. If I try to run `debug_test()` outside of a test file, then I
get an exception for trying to concatenate a nil value. This commit just
fixes the "no test found" short-circuit clause to handle empty string or
nil.