Set the Delve port dynamically
Currently the Delve port can be defined either randomly or hard-coded by
hand.

Hard-coded port allows to know which port to use in debug
configurations, but in practice with many Go projects, it would be
great to set the Delve port in debug configurations.

Luckily `nvim-dap` allows to set an `on_config` callback for an adapter,
which can be used to resolve the adapter settings when a debug session
starts. This PR implements this `nvim-dap` feature, which can be used
with the support of VSCode `launch.json` files to greatly enhance the developer
experience.

This potentially resolves https://github.com/leoluz/nvim-dap-go/issues/73