fix: do not add executable to adapter in launch request
The adapter would add the `executable` property to the adapter
regardless of whether the request was `launch` or `attach`. When
`attach` was used, it would attempt to start `dlv` on the same port that
it was attempting to configure.

This changes it so the executable is not added if `attach` is used and
adds some protection code to prevent the random port assignment from
being used with the attach configuration (where it wouldn't work
anyway).